changed parameter type of get all airport method in airport service

+++ src/main/java/com/aerotravel/flightticketbooking/services/AirportService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface AirportService {
     public abstract Page<Airport> getAllAirportsPaged(int pageNum);
     public abstract List<Airport> getAllAirports();
-    public abstract Airport getAirportById(Long airportId);
+    public abstract Airport getAirportById(Integer airportId);
     public abstract Airport saveAirport(Airport airport);
     public abstract void deleteAirport(Long airpportId);
 }
